backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Security-Too==5.3.2
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-RESTful==0.3.10
python-dotenv==1.0.0
mysqlclient==2.2.0
redis==5.0.1
PyJWT==2.8.0
Pillow==10.1.0
simple-websocket==1.0.0
gunicorn==21.2.0
python-engineio==4.8.0
python-socketio==5.10.0
webrtcvad==2.0.10
emoji==2.8.0
geopy==2.4.1
Werkzeug==3.0.1
itsdangerous==2.2.0
sqlalchemy==2.0.23
email-validator==2.1.0.post1
pymysql==1.1.0
